+++ sdk/src/main/java/ly/count/android/sdk/ModuleCrash.java
@@ -13,7 +13,7 @@ import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class ModuleCrash extends ModuleBase{
+class ModuleCrash extends ModuleBase{
     //native crash
     private static final String countlyFolderName = "Countly";
     private static final String countlyNativeCrashFolderName = "CrashDumps";